refactor(stats-aggregation-cli): add the following:
- Organize structure by separating commands and main logic into distinct directories.
- Utilize index file for command exports and introduce a dedicated runner file for executing main logic.
