+++ clis/stats-aggregation-cli/package.json
@@ -9,7 +9,8 @@
     "clean": "rm -rf .turbo node_modules",
     "lint": "eslint .",
     "lint:fix": "pnpm lint --fix",
-    "type-check": "tsc --noEmit"
+    "type-check": "tsc --noEmit",
+    "stats": "ts-node src/runner.ts"
   },
   "dependencies": {
     "@blobscan/db": "^0.0.1",
