+++ clis/stats-aggregation-cli/src/commands/index.ts
@@ -1,0 +1,2 @@
+export * from "./daily";
+export * from "./overall";
