+++ clis/stats-aggregation-cli/src/utils.ts
